fix deprecations after r25551: move cope Properties from demoshop into framework
git-svn-id: https://svn.exedio.com/cope/cope/trunk@25553 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/src/com/exedio/cope/ClusterSenderMulticast.java
@@ -34,7 +34,7 @@ final class ClusterSenderMulticast extends ClusterSender
 	{
 		super(properties);
 		this.address = properties.sendAddress;
-		this.port = properties.sendDestinationPort.intValue();
+		this.port = properties.sendDestinationPort;
 		this.socket = properties.newSendSocket();
 	}
